# Conflicts:
#	pulsar-broker/src/test/java/org/apache/pulsar/client/api/SimpleSchemaTest.java

+++ managed-ledger/src/main/java/org/apache/bookkeeper/mledger/ManagedLedgerInfo.java
@@ -24,7 +24,7 @@ import java.util.Map;
 @SuppressWarnings("checkstyle:javadoctype")
 public class ManagedLedgerInfo {
     /** Z-Node version. */
-    public int version;
+    public long version;
     public String creationDate;
     public String modificationDate;
 
@@ -42,7 +42,7 @@ public class ManagedLedgerInfo {
 
     public static class CursorInfo {
         /** Z-Node version. */
-        public int version;
+        public long version;
         public String creationDate;
         public String modificationDate;
